requirements.txt
altgraph==0.17
dataclasses-json==0.5.2
gitter-pkg==1.0.0
marshmallow==3.8.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.8
PyYAML==5.3.1
stringcase==1.2.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
